requirements.txt
torch
diffusers
transformers
accelerate
numpy
pillow